Get rid of the contacts table, use the roster
There's no point duplicating the contents of the roster in the database.
We now accept messages from anyone who sends them, and send reminders to
anyone who has subscribed to us. The web interface lists people who have
subscribed to us, and but lists messages for any jid that wrote to us.

Now we don't need to do the roster update dance at startup, and
%contacts stores only the last sent/received timestamps.